% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CADE.R
\docType{package}
\name{CADE}
\alias{CADE}
\alias{CADE-package}
\title{CADE: A package for using the cade algorithm to find outliers.}
\description{
The CADE package provides functions to find outliers and make uniform
distributions of each varaible type.
}

