% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uniformize_var.R
\name{uniform.default}
\alias{uniform.default}
\title{Default version of uniform}
\usage{
\method{uniform}{default}(x, len = length(x))
}
\arguments{
\item{x}{The vctor of data}

\item{len}{The length of the uniform set created}
}
\description{
Default version of uniform
}
\author{
Kenny Darrell
}
\keyword{cade}

