% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uniform.integer}
\alias{uniform.integer}
\title{Integer version of uniform}
\usage{
\method{uniform}{integer}(x, len = length(x))
}
\arguments{
\item{x}{The vctor of data}
}
\description{
Integer version of uniform
}
\author{
Kenny Darrell
}
\keyword{cade}

