% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uniformize_var.R
\name{uniform.numeric}
\alias{uniform.numeric}
\title{Numeric version of uniform}
\usage{
\method{uniform}{numeric}(x)
}
\arguments{
\item{x}{The vctor of data}
}
\description{
Numeric version of uniform
}
\author{
Kenny Darrell
}
\keyword{cade}

