% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uniform}
\alias{uniform}
\title{Create uniform distribution}
\usage{
uniform(x, ...)
}
\arguments{
\item{x}{The vctor of data}
}
\description{
This function will take an input and create a similar output that
is distributed over the same type.
}
\author{
Kenny Darrell
}
\keyword{cade}

