% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cade}
\alias{cade}
\title{The main CADE function}
\usage{
cade(df, prop = 1, numTree = 500, skip = c())
}
\arguments{
\item{df}{The data frame in question}

\item{prop}{The proportion of true to false cases}

\item{numTree}{The number of tree to create in the random forest}

\item{skip}{Fields that should not be included}
}
\description{
This function calls and runs CADE
}
\examples{
cade(cars)
}
\keyword{anonaly}
\keyword{detection}

