% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uniform.default}
\alias{uniform.default}
\title{Fefault version of uniform}
\usage{
\method{uniform}{default}(x, len = length(x))
}
\arguments{
\item{x}{The vctor of data}
}
\description{
Fefault version of uniform
}
\author{
Kenny Darrell
}
\keyword{cade}

