% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uniform.factor}
\alias{uniform.factor}
\title{Factor version of uniform}
\usage{
\method{uniform}{factor}(x, len = length(x))
}
\arguments{
\item{x}{The vctor of data}
}
\description{
Factor version of uniform
}
\author{
Kenny Darrell
}
\keyword{cade}

