% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uniform.numeric}
\alias{uniform.numeric}
\title{Numeric version of uniform}
\usage{
\method{uniform}{numeric}(x, len = length(x))
}
\arguments{
\item{x}{The vctor of data}
}
\description{
Numeric version of uniform
}
\author{
Kenny Darrell
}
\keyword{cade}

