% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uniformize_var.R
\name{uniform}
\alias{uniform}
\title{Create uniform distribution}
\usage{
uniform(x, len)
}
\arguments{
\item{x}{The vctor of data}

\item{len}{The length of the uniform set created}
}
\description{
This function will take an input and create a similar output that
is distributed over the same type.
}
\author{
Kenny Darrell
}
\keyword{cade}

